channel_manager/requirements.txt
pymongo
fastapi
pydantic
python-telegram
requests
uvicorn
